Changed behavior for tracks.txt
If -yt supplied and tracks file was not supplied, it tries to generate a track information by running the Tkinter GUI. The filename will be <YouTube Video ID>_tracks.txt. If it exists, (for some reason, if you try to run the split on the same video again) it will just read in the existing tracks file.