Big file problem solved!
Pydub wasn't capable of handling a file has more than 4GB of size for some reason. So, I've implemented subprocess routine to use external FFMpeg in this case. In fact, FFMpeg is being used by pydub if given codec is not LAME or OGG. So, there's nothing too much to worry about it implementing something dependent on ffmpeg.

Maybe I need to include FFMpeg into the package... But the script is supposed to be platform independent. So, just ask people to install FFMpeg instead!