Update tag_data to add song title & track number to tags
When splitting an album file the resulting song files
retain the same title as the original full album file.
e.g. if you had a full album file called "Pink Floyd - The Wall
(full).mp3", every track extracted would have the track title
"Pink Floyd - The Wall (full).mp3".

In addition, the track number is not included in tag_data, This results
in needing to manually update the ID3 title & track number tags with
only the track length (or listening to the track) as a clue to which
track is which.

Adding track title and track number to tag_data to fix this.

Also removed unused uuid & namedtuple imports.