Add playlist handling, tracklist extraction, and metadata extraction
If the provided link is a playlist, the process will be applied to each
item of this playlist.
If extract-track is provided, the tracklist will be extracted from youtube,
by looking in the video info, the description, and the first 10 comments.
If extract-metadata is provided, the name, artist, album, and year will be
extracted from the video, and used if not already provided.
A format can be provided, as storing youtube videos as wav is wasteful.
The main function has been adapted to be used outside of CLI